feat: Add Starter Kits mod
This commit introduces the Starter Kits mod, which allows players to choose from a variety of starter kits at the beginning of a game.  The mod includes features for creating, managing, and claiming kits.  This initial commit adds the core functionality, including network communication and a basic GUI.  Further improvements and features will be added in subsequent commits.